--- a/setup.cfg
+++ b/setup.cfg
@@ -36,7 +36,7 @@ install_requires =
     numpy>=1.16.5
     pandas>=1.1.1
     rxn-chem-utils>=1.0.4
-    rxn-utils>=1.4.0
+    rxn-utils>=1.4.1
     tabulate>=0.8.7
     xxhash>=2.0.0