--- a/src/rxn/reaction_preprocessing/augmenter.py
+++ b/src/rxn/reaction_preprocessing/augmenter.py
@@ -42,6 +42,7 @@ class Augmenter:
     Note: Unlike the other classes, which are memory-efficient, this one loads the
     whole data in a pandas DataFrame for processing.
     """
+
     def __init__(
         self, df: pd.DataFrame, reaction_column_name: str, fragment_bond: str = "."
     ):