Basically, this PR is solely about making the data processing script more efficient by avoiding loading of full CSV files into memory. Instead, the processing is done line-by-line.

This required a very different approach to the processing, where it was not possible to use `pandas.apply`. Instead, the code now relies on the `CsvIterator` approach from `rxn-utilities`.

All the classes were updated except the `Augmenter`.

Testing on the Pistachio dataset, we see the timings being slightly slower after the change, but the memory requirements are much smaller! For instance, at the standardization step, the difference is more than a factor of 100: ~110 MB vs ~35GB.
Differences in detail:

Before this PR:
```
    CPU time :                                   19233.39 sec.
    Max Memory :                                 47119 MB
    Average Memory :                             22810.98 MB
    Run time :                                   19343 sec.
    Turnaround time :                            19340 sec.
```

After this PR:
```
    CPU time :                                   15231.15 sec.
    Max Memory :                                 2165 MB
    Average Memory :                             397.43 MB
    Run time :                                   15314 sec.
    Turnaround time :                            15311 sec.
```

I.e., it is ~20% faster, and consumes 1/20 of the max memory, and ~1/57 of the average memory.

The generated test and validation splits are identical before and after (same md5sum). The train split is ordered differently; this is the only difference in behavior.